Inventory/UI
Created a basic inventory UI with slots.
Made the player able to open and close said inventory using the TAB key.
Started (concept in paint) on implementing textures like gun, medkit, rifle etc as dragable and usuable items.
Also changed the location of the healtbar as it did not appear for me (as it was outside of the canvas UI).