Created a new zombie(To later replace old one)
It walks towards the player and rotates to look at him.
It has a script holding its damage. Made a prefab of the zombie
In the player movement script there is a "OnCollisionStay2D" that checks if it collides with a zombie and then deals damage. There is a timer that blocks how often this can happen.
The spike hazard also got a "OnTriggerStay2D" and is also affected by the timer