Made a door that goes to level 2 in shop. Changed the "PayToOpen" script to be able to do more stuff
Renamed it to "Interactable"
It has 3 bools that can be turned on or off to switch what it should do when interacted with.
The things it can do its:
Take money to remove the object.
Go to a specifed level
Check for items and then go to a specified level.